requirements.txt
# Envyro-Core Requirements
# Python dependencies for the EnvyroAI neural engine

# Core Deep Learning
torch>=2.0.0
numpy>=1.24.0

# Database & Vector Search
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Password Hashing
bcrypt>=4.0.0

# Optional: For production embedding models
# sentence-transformers>=2.2.0
# transformers>=4.30.0

# Utilities
python-dotenv>=1.0.0

# Security & Encryption
cryptography>=42.0.0
PyJWT>=2.8.0

# Web Interface
flask>=2.0.0
flask-cors>=4.0.0
werkzeug>=2.0.0
